requirements.txt
flask
flask-jwt-extended
mysql-connector-python
python-dotenv
flask-cors
bcrypt
